medical-assistant/requirements.txt
torch
torchvision
transformers
langchain
ctransformers
gtts
matplotlib
Pillow
